Kaypro II boot disk images for GoTek drive
This zip file contains the images used on a Kaypro II computer with a GoTek drive. Unzip this files can copy to USB key, then insert into GoTek drive that is used for the Kaypro II and boot. You select the different images for different disks with the knob or buttons on your GoTek. If replacing Drive A, ensure your GoTek is set to S0 and if replacing Drive B then set to S1. This is the image I use and if you have question please search the internet for additional information as I only have this information\image that works on my computer.